// src/app/models/ricetta.ts
export class Ricetta {
    idRicetta: number;
    titolo: string;
    categoria: string;
    dieta: string;
    tempo: number;
    cottura: number;
    porzioni: string;
    descr: string;
    ingr: string
    prep: string;
    cons: string;
    img: string;
    autoreRic: string;
    video: string;

    // 
    // 
}